feat(runtime/rmi): reject overlapping VDEV ranges on the same PDEV
Cache each active VDEV's address ranges in the owning PDEV and use that
cache during smc_vdev_create() to reject new VDEVs whose ranges overlap
an existing VDEV in that PDEV. Storing ranges instead of vdev PAs in
the PDEV avoids locking other VDEV granules during create while still
enforcing same-PDEV range separation.

Track the claimed PDEV cache slot in each VDEV so smc_vdev_destroy()
can clear the cached ranges deterministically when the VDEV is torn
down.

+++ runtime/rmi/rtt.c
@@ -2114,7 +2114,7 @@ out_unlock_rec_rd:
 
 static bool addr_is_in_vdev_addr_range(struct vdev *vd, unsigned long addr)
 {
-	for (size_t i = 0; i < vd->num_addr_range; ++i) {
+	for (unsigned long i = 0; i < vd->num_addr_range; ++i) {
 		if ((addr >= vd->addr_range[i].base) &&
 		    (addr < vd->addr_range[i].top)) {
 			return true;
